restructuration repertoire (#424) (#426)
* restructuration repertoire

* correction
