+++ interview/client/src/pages/InterviewDeconnect.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
-import './InterviewStarted.css';
+import '../assets/css/InterviewStarted.css';
 
 function InterviewDeconnect() {
     return (
