+++ interview/package.json
@@ -1,0 +1,17 @@
+{
+  "dependencies": {
+    "concurrently": "^7.1.0"
+  },
+  "name": "interview",
+  "version": "1.0.0",
+  "main": "index.js",
+  "devDependencies": {},
+  "scripts": {
+    "client": "cd client && npm i && npm start",
+    "server": "cd server && npm i && npm start",
+    "dev": "concurrently \"npm run server\" \"npm run client\""
+  },
+  "author": "",
+  "license": "ISC",
+  "description": ""
+}
